# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.3.0
  description: |
    Namma Seva API (Phase 3). The chain is the source of truth; this API serves the indexed
    Postgres read model, pins metadata to IPFS for wallets to sign, authenticates wallets (SIWE)
    and citizens (phone OTP), and relays gasless citizen grievances through the ERC-2771 forwarder.
    Officials, auditors and contractors sign privileged transactions in their own wallets — there
    are no server-signed approve / reject / release routes.

    Amounts are decimal strings in the escrow's accounting unit (INR paise in LEDGER mode, wei in
    ESCROW mode). Coordinates are micro-degrees (`latE6`, `lngE6`).
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: {}
      required: [error]
...
    Role:
      type: string
      enum: [ADMIN, GOVT_OFFICIAL, AUDITOR, CONTRACTOR, CITIZEN, PUBLIC]
    SessionUser:
      type: object
      properties:
        id: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        walletAddress: { type: [string, "null"] }
        citizenHash: { type: [string, "null"] }
        wards:
          type: array
          items: { type: integer }
        preferredLang: { type: string }
        demo: { type: boolean }
      required: [id, role, roles, walletAddress, citizenHash, wards, preferredLang, demo]
    AuthTokens:
      type: object
      properties:
        accessToken: { type: string }
        expiresIn: { type: integer, description: Seconds until the access token expires. }
        user: { $ref: "#/components/schemas/SessionUser" }
      required: [accessToken, expiresIn, user]
...
    SiweVerifyRequest:
      type: object
      properties:
        message: { type: string, maxLength: 4000 }
        signature: { type: string, pattern: "^0x[0-9a-fA-F]+$" }
      required: [message, signature]
    OtpSendRequest:
      type: object
      properties:
        phone: { type: string, minLength: 10, maxLength: 16 }
        consentVersion:
          type: string
          minLength: 1
          maxLength: 32
          description: Version of the Privacy Notice the citizen accepted (DPDP Act 2023). Must equal the server's current version.
      required: [phone, consentVersion]
    OtpSendResponse:
      type: object
      properties:
        sent: { type: boolean }
        expiresIn: { type: integer }
        devCode: { type: string, description: "DEMO MODE ONLY — the code, shown on screen so anyone can try the citizen flow." }
      required: [sent, expiresIn]
    OtpVerifyRequest:
      type: object
      properties:
        phone: { type: string, minLength: 10, maxLength: 16 }
        code: { type: string, pattern: "^[0-9]{6}$" }
        consentVersion:
          type: string
          minLength: 1
          maxLength: 32
          description: Privacy Notice version accepted; recorded against the account on success.
        lang: { type: string, enum: [en, kn, ta, hi], description: Language the notice was shown in. }
      required: [phone, code, consentVersion]
...
    DemoLoginRequest:
      type: object
      properties:
        role: { $ref: "#/components/schemas/Role" }
      required: [role]

    # ─── Reference data ─────────────────────────────────────────────────
...
    GrievanceCategory:
      type: string
      enum: [QUALITY, DELAY, SAFETY, MISSING_WORK, CORRUPTION, OTHER]
    Grievance:
      type: object
      properties:
        id: { type: integer }
        projectId: { type: integer }
        category: { $ref: "#/components/schemas/GrievanceCategory" }
        status: { type: string, enum: [OPEN, ESCALATED, RESPONDED] }
        cid: { type: string }
        upvotes: { type: integer }
        action: { type: string, enum: [NONE, PAUSE_PROJECT, DISMISS] }
        responseCid: { type: [string, "null"] }
        respondedBy: { type: [string, "null"] }
        escalatedAt: { type: [string, "null"], format: date-time }
        respondBy: { type: [string, "null"], format: date-time }
        respondedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
        createdTx: { type: string }
        mine: { type: boolean }
        upvotedByMe: { type: boolean }
      required: [id, projectId, category, status, cid, upvotes, action, responseCid, respondedBy, escalatedAt, respondBy, respondedAt, createdAt, createdTx, mine, upvotedByMe]
    FileGrievanceRequest:
      type: object
      properties:
        projectId: { type: integer, minimum: 1 }
        category: { $ref: "#/components/schemas/GrievanceCategory" }
        text: { type: string, minLength: 10, maxLength: 2000 }
        lang: { type: string, enum: [en, kn, ta, hi] }
        photoCids:
          type: array
          maxItems: 4
          items: { type: string }
      required: [projectId, category, text]
    RelayJob:
      type: object
      properties:
        jobId: { type: string }
        kind: { type: string }
        status: { type: string, enum: [queued, registering, submitted, confirmed, failed] }
        txHash: { type: [string, "null"] }
        cid: { type: [string, "null"] }
        error: { type: [string, "null"] }
      required: [jobId, kind, status, txHash, cid, error]

    # ─── Tenders ────────────────────────────────────────────────────────
...
security: []
paths:
...
  /auth/siwe/verify:
    post:
      operationId: verifySiwe
      tags: [auth]
      summary: Verify a signed SIWE message; roles are read from NammaSevaAccess on-chain
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SiweVerifyRequest" }
      responses:
        "200":
          description: Signed in; also sets the httpOnly `ns_session` refresh cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: Invalid message or signature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/otp/send:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send a 6-digit login code to a citizen's phone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpSendRequest" }
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OtpSendResponse" }
        "429":
          description: Too many codes requested
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify the code and sign the citizen in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OtpVerifyRequest" }
      responses:
        "200":
          description: Signed in; also sets the `ns_session` cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: Wrong or expired code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /auth/refresh:
    post:
      operationId: refreshSession
      tags: [auth]
      summary: Rotate the refresh cookie and issue a new access token
      responses:
        "200":
          description: Refreshed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401":
          description: No valid session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security: [{ bearer: [] }]
      summary: Current user
      responses:
        "200":
          description: User
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          description: Not signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /auth/demo:
    post:
      operationId: demoLogin
      tags: [auth]
      summary: DEMO ONLY (NS_DEMO_MODE=true) — sign in as a role without a wallet or phone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DemoLoginRequest" }
      responses:
        "200":
          description: Demo session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "404":
          description: Demo mode is off
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ─── Reference ─────────────────────────────────────────────────────────
...
  /grievances:
    get:
      operationId: listGrievances
      tags: [grievances]
      parameters:
        - { name: projectId, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string, enum: [OPEN, ESCALATED, RESPONDED] } }
        - { name: mine, in: query, schema: { type: boolean } }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Grievances, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Grievance" }
    post:
      operationId: fileGrievance
      tags: [grievances]
      security: [{ bearer: [] }]
      summary: Citizen files a grievance; relayed gaslessly through the TrustedForwarder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileGrievanceRequest" }
      responses:
        "202":
          description: Queued for the relayer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelayJob" }
        "429":
          description: Daily limit reached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /grievances/{id}:
    get:
      operationId: getGrievance
      tags: [grievances]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Grievance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Grievance" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /grievances/{id}/upvote:
    post:
      operationId: upvoteGrievance
      tags: [grievances]
      security: [{ bearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "202":
          description: Queued for the relayer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelayJob" }
        "409":
          description: Already upvoted / own grievance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /relay/jobs/{jobId}:
    get:
      operationId: getRelayJob
      tags: [grievances]
      security: [{ bearer: [] }]
      parameters:
        - { name: jobId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Relay job status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelayJob" }
        "404":
          description: Unknown job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ─── Tenders ───────────────────────────────────────────────────────────
...